requirements.txt
aiohttp==3.7.3
pyserial==3.5
python-socketio==5.0.4
pylint==2.7.1
